Fix incorrect summarized size for containers and images. Ref #12

+++ src/main/scala/model/model.scala
@@ -19,7 +19,7 @@ case class DockerMetadata(connection: Connection,
                           version: Version,
                           containers: Seq[Container]) {
 
-  def totalContainersSize = bytesToSize(containers.map(c => c.SizeRootFs + c.SizeRw).sum)
+  def totalContainersSize = bytesToSize(containers.map(c => c.SizeRootFs.toLong + c.SizeRw.toLong).sum)
 }
 
 case class Info(Containers: Int,
